requirements.txt
Django==2.2.7
gunicorn==19.9.0
pytz==2019.3
sqlparse==0.3.0
whitenoise==4.1.4
dj-database-url==0.5.0
Pillow==6.0.0
psycopg2==2.8.2
django-heroku-0.3.1

